code/requirements.txt
cosmotech-run-orchestrator~=2.0.2
azure-storage-blob
CosmoTech-Acceleration-Library~=2.0.0
cosmotech-api~=5.0.0rc1
faker~=30.6.0
psycopg2-binary==2.9.*
numpy~=2.3.2
